service/requirements.txt
backports.ssl-match-hostname==3.5.0.1
click==6.7
Flask==0.12.1
Flask-SocketIO==2.8.6
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pyserial==3.3
python-engineio==1.5.1
python-socketio==1.7.4
requests==2.14.2
six==1.10.0
socketIO-client==0.7.2
virtualenv==15.1.0
websocket-client==0.40.0
Werkzeug==0.12.1
